requirements.txt
streamlit>=1.41.0
python-dotenv>=1.0.1
langchain>=0.3.0
langchain-openai>=0.2.0
